% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRcentage-package.R
\docType{package}
\name{pRcentage-package}
\alias{pRcentage}
\alias{pRcentage-package}
\title{pRcentage: A simple percentage formatting function}
\description{
A simple function to convert numeric values to percentage format.
}
\keyword{internal}
