% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pct.R
\name{to_pct}
\alias{to_pct}
\title{Convert to percentage format}
\usage{
to_pct(x)
}
\arguments{
\item{x}{A numeric element or vector to convert to percentage format}
}
\value{
A character vector with percentages as strings
}
\description{
This function takes a numeric vector and converts
each number into a percentage format by multiplying by 100 and appending a '\%'.
This function provides a simple way to format numeric values as percentages for presentation or reporting.
}
\examples{
library(magrittr)
# Convert numeric values to percentage format
c(0.1, 0.25, 0.45) \%>\% to_pct()
# Output: "10\%", "25\%", "45\%"

}
\author{
Sophie Hetche \email{s.hetche@students.uu.nl}
}
